Refactor settings mutation and query to shop module
